move: 개발에 사용되는 이미지는 src>assets폴더로 이동
