+++ src/app/(mobile)/loginInfo/setMyLanguage/page.tsx
@@ -5,8 +5,8 @@ import { useRouter } from "next/navigation";
 import { createClient } from "@/utils/supabase/client";
 import { useQuery } from "@tanstack/react-query";
 import { fetchLanguageName } from "@/api/firstSetting/fetchLanguageName";
-import caretleft from "@/../public/images/caret-left.svg";
-import check from "@/../public/images/check.svg";
+import caretleft from "@/assets/caret-left.svg";
+import check from "@/assets/Check.svg";
 import Image from "next/image";
 
 export default function SetMyLanguage() {
